pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.9.0.post1",
    "celery>=5.5.2",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.59",
    "langchain-docling>=0.2.0",
    "langchain-experimental>=0.3.4",
    "langchain-huggingface>=0.2.0",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.16",
    "langchain-text-splitters>=0.3.8",
    "langgraph==0.2.60",
    "langsmith>=0.3.42",
    "matplotlib>=3.10.3",
    "py-zerox>=0.0.7",
    "redis>=6.1.0",
    "typing-extensions>=4.13.2",
    "pdfminer>=20191125",
    "pdfminer-six>=20250506",
    "docling>=2.15.1",
    "pytesseract>=0.3.13",
    "pillow>=10.4.0",
    "pi-heif>=0.22.0",
    "unstructured-inference==0.2.4",
    "unstructured==0.5.6",
    "numpy>=1.26.4",
    "opencv-python>=4.6.0.66",
    "opencv-python-headless>=4.11.0.86",
    "autopep8>=2.3.2",
    "chromadb>=0.6.3",
    "langchain-chroma>=0.2.3",
]
